Made caretaker roles menue match the mockup
