+++ App/views/CaretakerRolesMenu/styles.js
@@ -1,17 +1,8 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-  button: {
-    margin: 10,
-    padding: 10
-  },
   container: {
     flex: 1,
-    backgroundColor: '#F5FCFF',
-  },
-  title: {
-    fontSize: 24,
-    textAlign: 'center',
-    margin: 15,
-  },
+    backgroundColor: '#FFF'
+  }
 });
